requirements.txt
youtube-dl
discord.py[voice]
uptime